## Eliminate double date parsing in Matrix temporal filters

### Problem

The Matrix component's temporal filtering system was experiencing severe performance degradation when processing time series data. Performance profiling revealed that the `useTemporalData` hook was spending **~1,200ms** parsing dates for a typical yearly dataset (8,760 rows), accounting for 66.8% of the total rendering time.

**Root cause**: Each date string was being parsed multiple times:
- Once in `extractValueFromDate()` 
- Again in `extractTemporalValue()`
- This resulted in 2N parse operations for N dates

### Performance Profile (Before)
```
RowFilter: 1,189.3 ms (66.8%)
├── parseFlexibleDate: 626.7 ms (35.2%)
└── parseFlexibleDate: 558.2 ms (31.4%)
Total parsing: ~1,185 ms
```

### Performance Profile (After)
```
RowFilter: ~80 ms (5-8%)
└── parseFlexibleDate: ~40 ms (2-3%)
Total parsing: ~40 ms
```

### Solution

Implemented a single-parse architecture in `useTemporalData`:
1. Parse each date string only once into a Date object
2. Extract all temporal values (day, month, hour, etc.) from the parsed Date
3. Cache the extracted values for reuse across different filter operations

### Key Changes

- Added `ParsedDateInfo` interface to store all temporal values
- Created a new `useMemo` block that parses dates once and extracts all values
- Modified value extraction logic to use pre-computed values instead of re-parsing

### Results

- **Performance improvement**: ~15x faster (from ~1,200ms to ~80ms)
- **User experience**: Filters should now open instantly with no/minimal UI freezing
- **Memory trade-off**: Minimal (~20% increase for significantly better performance)
- **Code impact**: Single file change, fully backward compatible



## Local tests:

Before:

<img width="1728" height="983" alt="Screenshot 2025-07-22 at 10 52 56" src="https://github.com/user-attachments/assets/4cfe9e21-f206-44b4-ba69-34c981cc54c1" />


After:

<img width="3456" height="1966" alt="image" src="https://github.com/user-attachments/assets/da4f2f18-9c97-454f-b086-2b154ea453dd" />