@hdinia 2025-07-23
Moved to : https://github.com/AntaresSimulatorTeam/AntaREST/pull/2606